Local host correction.
Because I am using Atom, and the atom-live-server Package, I had to make 
a change to dbhelper.js. Onl line 12 I had to remove the localhost and 
replace it with 127.0.0.1 to allow the app to work. atom-live-server 
uses fills in 127.0.0.1 to the browser buy default so nothing happens if 
the code is set to localhost